Fix rename bug of build.sh Docker

+++ build.sh
@@ -45,24 +45,18 @@ cat <<EOF >"$BUILD_SCRIPT"
         --extra-version="VVCEasy"
     make -j\$(nproc) V=1
     make install install-doc
-EOF
 
-cat <<EOF >"$RENAME_VVCEASY"
-    if [[ "${TARGET}" == win* ]]; then
-        mv /ffbuild/prefix/bin/ffmpeg.exe /ffbuild/prefix/bin/ffmpeg_vvceasy.exe
-        mv /ffbuild/prefix/bin/ffprobe.exe /ffbuild/prefix/bin/ffprobe_vvceasy.exe
-        mv /ffbuild/prefix/bin/ffplay.exe /ffbuild/prefix/bin/ffplay_vvceasy.exe
-    else
-        mv /ffbuild/prefix/bin/ffmpeg /ffbuild/prefix/bin/ffmpeg_vvceasy
-        mv /ffbuild/prefix/bin/ffprobe /ffbuild/prefix/bin/ffprobe_vvceasy
-        mv /ffbuild/prefix/bin/ffplay /ffbuild/prefix/bin/ffplay_vvceasy
-    fi
+    [[ -f /ffbuild/prefix/bin/ffmpeg.exe ]] && mv /ffbuild/prefix/bin/ffmpeg.exe /ffbuild/prefix/bin/ffmpeg_vvceasy.exe
+    [[ -f /ffbuild/prefix/bin/ffprobe.exe ]] && mv /ffbuild/prefix/bin/ffprobe.exe /ffbuild/prefix/bin/ffprobe_vvceasy.exe
+    [[ -f /ffbuild/prefix/bin/ffplay.exe ]] && mv /ffbuild/prefix/bin/ffplay.exe /ffbuild/prefix/bin/ffplay_vvceasy.exe
+    [[ -f /ffbuild/prefix/bin/ffmpeg ]] && mv /ffbuild/prefix/bin/ffmpeg /ffbuild/prefix/bin/ffmpeg_vvceasy
+    [[ -f /ffbuild/prefix/bin/ffprobe ]] && mv /ffbuild/prefix/bin/ffprobe /ffbuild/prefix/bin/ffprobe_vvceasy
+    [[ -f /ffbuild/prefix/bin/ffplay ]] && mv /ffbuild/prefix/bin/ffplay /ffbuild/prefix/bin/ffplay_vvceasy
 EOF
 
 [[ -t 1 ]] && TTY_ARG="-t" || TTY_ARG=""
 
 docker run --rm -i $TTY_ARG "${UIDARGS[@]}" -v "$PWD/ffbuild":/ffbuild -v "$BUILD_SCRIPT":/build.sh "$IMAGE" bash /build.sh
-docker run --rm -i $TTY_ARG "${UIDARGS[@]}" -v "$PWD/ffbuild":/ffbuild -v "$RENAME_VVCEASY" "$IMAGE"
 
 mkdir -p artifacts
 ARTIFACTS_PATH="$PWD/artifacts"
